Make more correct

+++ src/Access.jl
@@ -9,4 +9,4 @@
 """ Access
 export Access, READ_ONLY, READ_WRITE, CREATE
 
-Base.hash(acc::Access, u::UInt) = hash(hash(convert(UInt, acc), u), UInt(0x0bec3b3e))
+Base.hash(acc::Access, u::UInt) = hash(hash(convert(UInt, acc), u), UInt(0x0bec3b3e))::UInt
